Another workaround related to case-sensitivity when a smil file refers to SMIL10.dtd (note the caps in the file name), the sample book included smil10.dtd (note the lower-case file name). In the smil files e.g. icth0011.smil the reference is to SMIL10.dtd. Again there seems to be a difference in behaviour of the SAX parser between Android and Java - on Android the lookup seems to be case-insensitive, on my workstation it's case-sensitive.
The DAISY 202 specification implies the filename lookup should be case sensitive. 

Anyway, so I can continue testing I'll check in a copy of the file that matches the name specified in the various 
smil files from the sample book (from www.daisy.org).

As a thought it'd be good to have a definitive DAISY checker, similar to http://code.google.com/p/epubcheck/