Another interim release. This one contains a mix of the various localised strings because I localized a previously unlocalized item (Up 1 level). I haven't updated Norwegian as it needs many more strings updating first, and I prefer to have a full update rather than tinkering with online translation tools for each missing item.
There's also some new code to allow reading to start from where we left-off last time we read this book. I'm not full convinced with my changes so there may be latent, quirky and annoying behaviours. 

I have also noticed the handling of missing audio files could be improved as users are unable to navigate past the problem. I will try to improve this in the next few commits.