I've completed the migration to use integers to represent the timings for the audio clips / segments.
In the process, I ended up changing the behaviour to accept timings without a decimal point. This was a side-effect of using DecimalFormat("###.###") which returns 0 for the input of 0.0 - since the overall change seems a slight improvement I've decided to live with this behaviour even though it means the code doesn't quite meet the DAISY 2.02 specification.