Added a custom -pre-build step to the main build.xml used by ant build that copies test DAISY content to the device.
The test DAISY content is intended for some of the unit tests. For some as-yet unknown reason I have to specify this -pre-build in the trunk rather than in a test project. Android test-projects first call the target (app's) build.xml which means -pre-build seems to be already defined by the time the testproject's build.xml is actually processed. Anyhow, the current approach seems adequate to enable me to continue working on getting all the tests to run cleanly on the device as well as in plain junit.